Add initial image processing functionality using Together AI API.
