+++ packages/llm/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
+    "together-ai": "^0.6.0-alpha.4",
     "xml2js": "^0.6.2"
   }
 }
