+++ packages/llm/prompts/lead-validate.ts
@@ -8,7 +8,6 @@ export const validateLeadWithInput = async ({
   leadMagnet: LeadMagnet;
   promptInput: any;
 }) => {
-  console.log({ leadMagnet });
   if (leadMagnet?.output === "image") {
     return await callImageLLM(leadMagnet, promptInput);
   }
